pyproject.toml
[tool.poetry]
name = "azure-api-functionapp-poc"
version = "1.0.0"
description = ""
authors = ["Mykhailo Chalyi <mike@chaliy.name>"]

[tool.poetry.dependencies]
python = "~3.9"
fastapi = "^0.71.0"
uvicorn = "^0.17.0"
azure-functions = "^1.8.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
